+++ src/components/Badge/tests/Badge.test.stories.tsx
@@ -37,3 +37,18 @@ export const onClick: StoryObj<{ handleClick: ReturnType<typeof fn> }> = {
 		expect(args.handleClick).toHaveBeenCalledWith(expect.objectContaining({ target: button }));
 	},
 };
+
+export const className: StoryObj = {
+	name: "className, attributes",
+	render: () => (
+		<div data-testid="root">
+			<Badge color="primary" className="test-classname" attributes={{ id: "test-id" }} />
+		</div>
+	),
+	play: async ({ canvas }) => {
+		const root = canvas.getByTestId("root").firstChild;
+
+		expect(root).toHaveClass("test-classname");
+		expect(root).toHaveAttribute("id", "test-id");
+	},
+};