+++ src/components/Card/Card.types.ts
@@ -3,7 +3,7 @@ import type { ActionableProps } from "components/Actionable";
 import type { ViewProps } from "components/View";
 import type * as G from "types/global";
 
-export type Props<TagName extends keyof JSX.IntrinsicElements = "div"> = {
+export type Props<TagName extends keyof React.JSX.IntrinsicElements = "div"> = {
 	padding?: G.Responsive<number>;
 	bleed?: G.Responsive<number>;
 	selected?: boolean;